// apps/coinet-platform/src/l13/contracts/l13-constitutional-types.ts
export enum L13ForbiddenAction {
  INVENT_MISSING_SUPPORT = 'INVENT_MISSING_SUPPORT',
  HIDE_CONTRADICTION = 'HIDE_CONTRADICTION',
  IGNORE_RESTRICTION = 'IGNORE_RESTRICTION',
  OVERRIDE_CONFIDENCE_CAP = 'OVERRIDE_CONFIDENCE_CAP',
  REBUILD_SCENARIO = 'REBUILD_SCENARIO',
  REBUILD_SCORE = 'REBUILD_SCORE',
  REBUILD_HYPOTHESIS = 'REBUILD_HYPOTHESIS',
  REBUILD_SEQUENCE = 'REBUILD_SEQUENCE',
  REBUILD_REGIME = 'REBUILD_REGIME',
  CREATE_NEW_SCENARIO = 'CREATE_NEW_SCENARIO',
  CREATE_NEW_HYPOTHESIS = 'CREATE_NEW_HYPOTHESIS',
  COMPUTE_SCORE_LOCALLY = 'COMPUTE_SCORE_LOCALLY',
  TREAT_SCENARIO_CONFIDENCE_AS_PROBABILITY =
    'TREAT_SCENARIO_CONFIDENCE_AS_PROBABILITY',
  EMIT_BUY_INSTRUCTION = 'EMIT_BUY_INSTRUCTION',
  EMIT_SELL_INSTRUCTION = 'EMIT_SELL_INSTRUCTION',
  EMIT_HOLD_INSTRUCTION = 'EMIT_HOLD_INSTRUCTION',
  EMIT_AVOID_INSTRUCTION = 'EMIT_AVOID_INSTRUCTION',
  EMIT_LEVERAGE_INSTRUCTION = 'EMIT_LEVERAGE_INSTRUCTION',
  EMIT_POSITION_SIZE_INSTRUCTION = 'EMIT_POSITION_SIZE_INSTRUCTION',
  EMIT_ENTRY_EXIT_INSTRUCTION = 'EMIT_ENTRY_EXIT_INSTRUCTION',
  CLAIM_GUARANTEE = 'CLAIM_GUARANTEE',
  CLAIM_CERTAINTY_UNSUPPORTED = 'CLAIM_CERTAINTY_UNSUPPORTED',
  CALL_SCENARIO_WINNER = 'CALL_SCENARIO_WINNER',
  CALL_HYPOTHESIS_FINAL_TRUTH = 'CALL_HYPOTHESIS_FINAL_TRUTH',
  SAY_SCORE_MEANS_RECOMMENDATION = 'SAY_SCORE_MEANS_RECOMMENDATION',
  PRETEND_MISSING_DATA_COMPLETE = 'PRETEND_MISSING_DATA_COMPLETE',
  OMIT_ACTIVE_INVALIDATION = 'OMIT_ACTIVE_INVALIDATION',
  OMIT_REQUIRED_TRIGGER = 'OMIT_REQUIRED_TRIGGER',
  USE_RAW_LOWER_LAYER_BYPASS = 'USE_RAW_LOWER_LAYER_BYPASS',
  OUTPUT_UNGROUNDED_CLAIM = 'OUTPUT_UNGROUNDED_CLAIM',
}
// ...
export enum L13ViolationSeverity {
  CRITICAL = 'CRITICAL',
  ERROR = 'ERROR',
  WARNING = 'WARNING',
}